worker/requirements.txt
#
####### How to install requirements #######
#
# pip install -r requirements.txt
#
###########################################
#

overpy
paho-mqtt
requests

